+++ Assets/Scripts/UI/Popup/UIModeSelect.cs
@@ -1,6 +1,7 @@
 using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
+using UnityEngine.Rendering;
 using UnityEngine.SceneManagement;
 using UnityEngine.XR.ARFoundation;
 
@@ -31,28 +32,20 @@ public class UIModeSelect : UIPopup
 
     private void OnClickNormalMode()
     {
-        Managers.UI.ClosePopupUI();
-        Managers.UI.ShowPopupUI<UIConfirmDialog>().SetDialog(
-            MoveToInstallBowlStep,
-            "안내",
-            "화면을 돌리면서 평면을 인식한 후\n바닥을 터치하여 그릇을 설치하세요.",
-            "확인");
-    }
+        SceneManager.LoadScene("InGame");
 
-    private void MoveToInstallBowlStep()
-    {
-        Managers.UI.ShowPopupUI<UISetPlatePosition>();
+        LoaderUtility.Deinitialize();
+        LoaderUtility.Initialize();
     }
 
     private void OnClickAttackMode()
     {
+        SceneManager.LoadScene("InGame");
+
+        LoaderUtility.Deinitialize();
+        LoaderUtility.Initialize();
+
         timeAttackMode = true;
-        Managers.UI.ClosePopupUI();
-        Managers.UI.ShowPopupUI<UIConfirmDialog>().SetDialog(
-            MoveToInstallBowlStep,
-            "안내",
-            "화면을 돌리면서 평면을 인식한 후\n바닥을 터치하여 그릇을 설치하세요.",
-            "확인");
         // 타임어택일때 불값으로 UIIngame에서 true인지 확인하고
         // 타이머를 초기화 하는 로직이 필요할듯
     }
